Q4_work/webhook-api/requirements.txt
streamlit 
fastapi 
uvicorn 
twilio 
python-dotenv 
requests 
openai